stable extraction of blog code in theme-common
