+++ packages/docusaurus-plugin-content-blog/src/client/index.tsx
@@ -1,0 +1,24 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export {
+  BlogPostProvider,
+  type BlogPostContextValue,
+  useBlogPost,
+  useBlogMetadata,
+} from './contexts';
+
+export {
+  useBlogListPageStructuredData,
+  useBlogPostStructuredData,
+} from './structuredDataUtils';
+
+export {
+  BlogSidebarItemList,
+  groupBlogSidebarItemsByYear,
+  useVisibleBlogSidebarItems,
+} from './sidebarUtils';
