+++ packages/docusaurus-theme-common/src/index.ts
@@ -39,11 +39,6 @@ export {
   filterDocCardListItems,
 } from './utils/docsUtils';
 
-export {
-  useBlogListPageStructuredData,
-  useBlogPostStructuredData,
-} from './utils/structuredDataUtils';
-
 export {usePluralForm} from './utils/usePluralForm';
 
 export {useCollapsible, Collapsible} from './components/Collapsible';
