+++ packages/docusaurus-theme-common/src/types.d.ts
@@ -6,6 +6,5 @@
  */
 
 /// <reference types="@docusaurus/module-type-aliases" />
-/// <reference types="@docusaurus/plugin-content-blog" />
 /// <reference types="@docusaurus/plugin-content-docs" />
 /// <reference types="@docusaurus/plugin-content-pages" />
